fix(GPUDevice): waitDevice idle before destroy

+++ src/core/gpu/shaders/square.hlsl
@@ -1,5 +1,5 @@
-[[vk::binding(0, 0)]] RWStructuredBuffer<int> InBuffer;
-[[vk::binding(1, 0)]] RWStructuredBuffer<int> OutBuffer;
+[[vk::binding(0, 0)]] RWStructuredBuffer<float> InBuffer;
+[[vk::binding(1, 0)]] RWStructuredBuffer<float> OutBuffer;
 
 [numthreads(1, 1, 1)]
 void Main(uint3 DTid : SV_DispatchThreadID)
